pyproject.toml
[tool.poetry]
name = "craft"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
pandas = "^2.2.2"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
greenlet = "^3.0.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
